[refactor] updated routes structure
- use async, await
- add auth app
- remove dotenv
