+++ docker-compose.yml
@@ -1,5 +1,5 @@
 services:
-  mongodb:
+  mongo:
     image: mongo
     restart: always
     ports:
