+++ package.json
@@ -16,7 +16,6 @@
   "dependencies": {
     "bcryptjs": "^2.4.3",
     "crypto-random-string": "^3.0.0",
-    "dotenv": "^7.0.0",
     "express": "^4.16.4",
     "express-flash": "0.0.2",
     "express-session": "^1.15.6",
