+++ src/utils/SendMail.js
@@ -1,12 +1,12 @@
 const nodemailer = require('nodemailer');
-const Room = require('../apps/room/Room.model');
+const Room = require('../apps/room/models');
 const Email_Id = process.env.Email_Id;
 const Email_Pass = process.env.Email_Pass;
 
 module.exports = {
     verify: (email, token, link) => {
         return new Promise((resolve, reject) => {
-            let transporter = nodemailer.createTransport({
+            const transporter = nodemailer.createTransport({
                 service: 'gmail',
                 auth: {
                     user: Email_Id,
@@ -17,7 +17,7 @@ module.exports = {
             let newlink = `${link}/user/verify/${email}/${token}`;
 
             let mailOptions = {
-                from: 'Admin myroomy <myroomy.admin@gmail.com>',
+                from: 'Admin roomy <roomy.admin@gmail.com>',
                 to: email,
                 subject: 'Account Verify',
                 text: `Verify Account Link <a href='${newlink}>Verify</a>'`,
@@ -61,26 +61,15 @@ module.exports = {
           `,
             };
 
-            //   transporter.sendMail(mailOptions, (error, info) => {
-            //     if (error) {
-            //       console.log(error);
-            //       console.log("Mail not sended");
-            //       reject(error);
-            //     } else {
-            //       Room.findOneAndUpdate(
-            //         { _id: id },
-            //         {
-            //           $set: {
-            //             status: false,
-            //           },
-            //         }
-            //       ).then((result) => {
-            //         resolve(result);
-            //         console.log("Update after email", result);
-            //       });
-            //       console.log("Mail Sended");
-            //     }
-            //   });
+            transporter.sendMail(mailOptions, (error, info) => {
+                if (error) {
+                    console.log(error);
+                    console.log('Mail not sended');
+                    reject(error);
+                }
+                resolve(result);
+                console.log('Mail Sended');
+            });
         });
     },
 };
