+++ src/views/user/dashboard.pug
@@ -28,4 +28,4 @@ block content
 						span Status
 						span #{status}
 			.editLink
-				a(href="/user/edit").btn.btn-outline-primary.btn-sm Edit Detail
+				a(href="/user/update").btn.btn-outline-primary.btn-sm Edit Detail
