+++ src/views/user/login.pug
@@ -4,7 +4,7 @@ block content
 	div.loginContainer.p-3.bg-white.rounded
 		h4.text-center.text-muted.mb-4.mt-3 Login
 		div.row
-			form(action='/user/auth' method='post').col-10.col-md-6.mx-auto.pl-0
+			form(action='/login' method='post').col-10.col-md-6.mx-auto.pl-0
 				div.form-group
 					label(for="exampleInputEmail1") UserID
 					input.ml-3(type="text" name='userid' class="form-control" id="exampleInputEmail1" aria-describedby="emailHelp" placeholder="Enter userid" required)
